Time: 0 ms (100%), Space: 6.4 MB (9.3%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -6,7 +6,7 @@ public:
             negative = true;
             x *= -1;
         }
-        long long reversed = 0;
+        long reversed = 0;
         while(x > 0){
             reversed = (reversed * 10) + (x % 10);
             x /= 10;
